requirements.txt
streamlit
torch
transformers
sentencepiece
openai-whisper
ffmpeg-python
